handle NOSCRIPT errors properly
All the scripts are loaded on application startup and the SHAs are
used there after. Even though this would work most of the time, we
can't assume the script would be always cached. Redis server restart
could wipe the script cache.

Fallback to eval command if the evalsha failed. The eval command
caches the script, so subsequent evalsha should succeed.